Improve communication with daemon process
